Add dynamic trust anchor for Certificate Store
Add an optional callback that allows for a certificate store to search
among known certs that don't fit in RAM (in the ta[] array). This allows
for hundreds of certs to be used even on the ESP8266 as only a single one
needs to be in RAM at any time instead of the whole chain.