Free RAM by not using jump tables
Jump tables using the current ESP8266 toolchain are placed in RODATA,
which is copied into RAM on startup.  Until GCC is built with the
-DJUMP_TABLES_IN_TEXT=1 macro, disable jump tables here and free
about 1.3KB of RAM for all apps.  With the tight memory constraints
of doing SSL on this chip, every byte helps and the small performance
degradation should not be bad compared to the long-running RSA/EC
codes.