Add STACK_PROXY_* macros to allocate some locals
Allow an application to get a small stack from the heap at the beginning of use
and use that space for allocating large variables.  This lets apps which do not
use the SSL code to not need larger stacks than they'd otherwise have to.

Still WIP, but runs to completion now on bearssl.org connections.